入力処理関係作成＆完了
InputEnumで取得するキーをリスト化
InputStateでビットにまとめて保存と取得を補助
InputFrameでフレーム別で保存とStateの一元管理
InputManagerでStateの変更と比較を担当

to_index補助関数をUtilityにしたためDrawResourcesクラスでのスコープ修正
その他暗黙変換があった場所を最適化

Input自体はまだ利用できていないためこれから利用できるようにwindowInterface等を改良

動作確認済み

+++ RendererInterface/DirectX/DirectXUpdater.h
@@ -87,7 +87,7 @@ namespace render {
 
 			//@breif	== フレームリソースサイズ保存変数 ==
 			//@details	描画に使うフレームリソースサイズを保存
-			UINT frame_resouse_size{};
+			size_t frame_resouse_size{};
 
 			//@brief	== フレームリソースサイクル管理用変数 ==
 			//@details	使用するフレームリソースインデックスを保存
