+++ RendererInterface/DirectX/DirectXobject/IndexBuffer.cpp
@@ -12,7 +12,7 @@ using namespace render::dx12::object;
 //@param	data	初期設定データ構造体に設定する先頭ポインター
 //@param	size	初期設定データ構造体に設定するメモリサイズ
 //@return	作成した構造体
-[[nodiscard]] render::dx12::desc::StaticBufferCreateDesc IndexBuffer::create_static_buffer_desc(const void* data, UINT64 size) {
+[[nodiscard]] render::dx12::desc::StaticBufferCreateDesc IndexBuffer::create_static_buffer_desc(const void* data, UINT size) {
 	
 	//	IndexBuffer用リソース作成
 	render::dx12::desc::StaticBufferCreateDesc desc{};
